+++ GraknClient.java
@@ -31,13 +31,13 @@ import grakn.client.concept.RelationType;
 import grakn.client.concept.Role;
 import grakn.client.concept.Rule;
 import grakn.client.concept.SchemaConcept;
-import grakn.client.concept.answer.Answer;
-import grakn.client.concept.answer.AnswerGroup;
-import grakn.client.concept.answer.ConceptList;
-import grakn.client.concept.answer.ConceptMap;
-import grakn.client.concept.answer.ConceptSet;
-import grakn.client.concept.answer.ConceptSetMeasure;
-import grakn.client.concept.answer.Numeric;
+import grakn.client.answer.Answer;
+import grakn.client.answer.AnswerGroup;
+import grakn.client.answer.ConceptList;
+import grakn.client.answer.ConceptMap;
+import grakn.client.answer.ConceptSet;
+import grakn.client.answer.ConceptSetMeasure;
+import grakn.client.answer.Numeric;
 import grakn.client.exception.GraknClientException;
 import grakn.client.rpc.RequestBuilder;
 import grakn.client.rpc.ResponseReader;
