+++ answer/ConceptList.java
@@ -16,29 +16,26 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package grakn.client.concept.answer;
+
+package grakn.client.answer;
 
 import grakn.client.concept.ConceptId;
 
 import java.util.Collections;
-import java.util.Set;
+import java.util.List;
 
 /**
- * A type of Answer object that contains a Set.
+ * A type of Answer object that contains a List of Concepts.
  */
-public class ConceptSet extends Answer {
+public class ConceptList extends Answer {
 
-    // TODO: change to store Set<Concept> once we are able to construct Concept without a database look up
-    private final Set<ConceptId> set;
+    // TODO: change to store List<Concept> once we are able to construct Concept without a database look up
+    private final List<ConceptId> list;
     private final Explanation explanation;
 
-    public ConceptSet(Set<ConceptId> set) {
-        this(set, new Explanation());
-    }
-
-    public ConceptSet(Set<ConceptId> set, Explanation explanation) {
-        this.set = Collections.unmodifiableSet(set);
-        this.explanation = explanation;
+    public ConceptList(List<ConceptId> list) {
+        this.list = Collections.unmodifiableList(list);
+        this.explanation = new Explanation();
     }
 
     @Override
@@ -46,20 +43,20 @@ public class ConceptSet extends Answer {
         return explanation;
     }
 
-    public Set<ConceptId> set() {
-        return set;
+    public List<ConceptId> list() {
+        return list;
     }
 
     @Override
     public boolean equals(Object obj) {
         if (obj == this) return true;
         if (obj == null || getClass() != obj.getClass()) return false;
-        ConceptSet a2 = (ConceptSet) obj;
-        return this.set.equals(a2.set);
+        ConceptList a2 = (ConceptList) obj;
+        return this.list.equals(a2.list);
     }
 
     @Override
     public int hashCode() {
-        return set.hashCode();
+        return list.hashCode();
     }
 }
