+++ answer/ConceptSetMeasure.java
@@ -17,46 +17,46 @@
  * under the License.
  */
 
-package grakn.client.concept.answer;
+package grakn.client.answer;
 
 import grakn.client.concept.ConceptId;
 
-import java.util.Collections;
-import java.util.List;
+import java.util.Set;
 
 /**
- * A type of Answer object that contains a List of Concepts.
+ * A type of Answer object that contains a Set and Number, by extending RemoteConceptSet.
  */
-public class ConceptList extends Answer {
+public class ConceptSetMeasure extends ConceptSet {
 
-    // TODO: change to store List<Concept> once we are able to construct Concept without a database look up
-    private final List<ConceptId> list;
-    private final Explanation explanation;
+    private final Number measurement;
 
-    public ConceptList(List<ConceptId> list) {
-        this.list = Collections.unmodifiableList(list);
-        this.explanation = new Explanation();
+    public ConceptSetMeasure(Set<ConceptId> set, Number measurement) {
+        this(set, measurement, new Explanation());
     }
 
-    @Override
-    public Explanation explanation() {
-        return explanation;
+    public ConceptSetMeasure(Set<ConceptId> set, Number measurement, Explanation explanation) {
+        super(set, explanation);
+        this.measurement = measurement;
     }
 
-    public List<ConceptId> list() {
-        return list;
+    public Number measurement() {
+        return measurement;
     }
 
     @Override
     public boolean equals(Object obj) {
         if (obj == this) return true;
         if (obj == null || getClass() != obj.getClass()) return false;
-        ConceptList a2 = (ConceptList) obj;
-        return this.list.equals(a2.list);
+        ConceptSetMeasure a2 = (ConceptSetMeasure) obj;
+        return this.set().equals(a2.set())
+                && measurement.toString().equals(a2.measurement.toString());
     }
 
     @Override
     public int hashCode() {
-        return list.hashCode();
+        int hash = super.hashCode();
+        hash = 31 * hash + measurement.hashCode();
+
+        return hash;
     }
 }
