+++ test/deployment/src/test/java/application/MavenApplicationTest.java
@@ -1,7 +1,7 @@
 package application;
 
 import grakn.client.GraknClient;
-import grakn.client.concept.answer.ConceptMap;
+import grakn.client.answer.ConceptMap;
 import graql.lang.Graql;
 import org.junit.Test;
